requirements.txt
certifi==2019.11.28
chardet==3.0.4
Click==7.0
dnspython==1.16.0
Flask==1.1.1
Flask-MySQL==1.5.1
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
mysql-connector-python==8.0.19
nba-api==1.1.8
numpy==1.18.1
pandas==1.0.1
passlib==1.7.2
pkg-resources==0.0.0
protobuf==3.6.1
PyMySQL==0.9.3
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
sha256==0.1
six==1.14.0
urllib3==1.25.8
uWSGI==2.0.18
Werkzeug==1.0.0
